Nab Integer->Long

+++ nab/src/org/labkey/nab/NabDataHandler.java
@@ -115,7 +115,7 @@ public abstract class NabDataHandler extends DilutionDataHandler
     protected void importRows(ExpData data, ExpRun run, ExpProtocol protocol, DataIteratorBuilder rawData, User user) throws ExperimentException
     {
         Map<Integer, String> cutoffFormats = getCutoffFormats(protocol, run);
-        Map<String, Pair<Integer, String>> wellGroupNameToNabSpecimen = new HashMap<>();
+        Map<String, Pair<Long, String>> wellGroupNameToNabSpecimen = new HashMap<>();
 
         populateDilutionStats(data, run, protocol, rawData, wellGroupNameToNabSpecimen);
         populateWellData(protocol, run, user, cutoffFormats, wellGroupNameToNabSpecimen);
@@ -125,7 +125,7 @@ public abstract class NabDataHandler extends DilutionDataHandler
      * Populates cutoff and AUC information from the passed in raw data
      */
     public void populateDilutionStats(ExpData data, ExpRun run, ExpProtocol protocol, DataIteratorBuilder rawData,
-                                      Map<String, Pair<Integer, String>> wellgroupNameToNabSpecimen) throws ExperimentException
+                                      Map<String, Pair<Long, String>> wellgroupNameToNabSpecimen) throws ExperimentException
     {
         _populateDilutionStats(data, run, protocol, rawData, wellgroupNameToNabSpecimen, true, Collections.emptyList(), Collections.emptyList());
     }
@@ -148,7 +148,7 @@ public abstract class NabDataHandler extends DilutionDataHandler
      * @param cutoffRows if commitData is false, then cutoff data will be returned in this collection
      */
     private void _populateDilutionStats(ExpData data, ExpRun run, ExpProtocol protocol, DataIteratorBuilder rawData,
-                                        Map<String, Pair<Integer, String>> wellGroupNameToNabSpecimen, boolean commitData,
+                                        Map<String, Pair<Long, String>> wellGroupNameToNabSpecimen, boolean commitData,
                                         List<Map<String, Object>> specimenRows, List<Map<String, Object>> cutoffRows) throws ExperimentException
     {
         Container container = run.getContainer();
@@ -206,7 +206,7 @@ public abstract class NabDataHandler extends DilutionDataHandler
                 nabSpecimenEntries.put("VirusLsid", createVirusWellGroupLsid(data, virusWellGroupName));
                 nabSpecimenEntries.put(FIT_PARAMETERS_PROPERTY_NAME, group.get(FIT_PARAMETERS_PROPERTY_NAME));
 
-                int nabRowid = 0;
+                long nabRowid = 0;
                 if (commitData)
                 {
                     nabRowid = NabManager.get().insertNabSpecimenRow(null, nabSpecimenEntries);
